Adding materials
LBCN_epoch: allowing to epoch and baseline correct with more flexibility
than SPM, as it can use the RT instead of the event onset while keeping
a baseline relative to the onset.
utils/get_events_diod: detect the onset, duration and offset of each
event based on the diod signal. Can be used to create events.mat later
on.